improve redirectios and clean URLs

+++ classes/Utils.php
@@ -92,13 +92,19 @@ class Utils
 	 * Check if a user exists and get the profile
 	 *
 	 * @author salvipascual
+	 * @param String $niddle: Can be an email, @username or ID
 	 * @return object|boolean
 	 */
-	public static function getPerson($email)
+	public static function getPerson($niddle)
 	{
+		// select where condition from @username, email or ID
+		if(substr($niddle, 0, 1) === "@") $where = "LOWER(username)=LOWER('$niddle')";
+		elseif(strpos($niddle, '@') !== false) $where = "LOWER(email)=LOWER('$niddle')";
+		else $where = "id='$niddle'";
+
 		// get the person via email
-		$person = Connection::query("SELECT * FROM person WHERE LOWER(email)=LOWER('$email')");
-		return $person ? $person[0] : false;
+		$person = Connection::query("SELECT * FROM person WHERE $where");
+		return $person ? Social::prepareUserProfile($person[0]) : false;
 	}
 
 	/**
@@ -122,49 +128,6 @@ class Utils
 		} while($exist);
 
 		return $username;
-	}
-
-	/**
-	 * Get the email from an username
-	 *
-	 * @author salvipascual
-	 * @param String $username
-	 * @return stdClass $person or false
-	 */
-	public static function getPersonFromUsername($username)
-	{
-		// do not try empty inputs
-		if(empty($username)) return false;
-
-		// remove the @ symbol
-		$username = str_replace("@", "", $username);
-
-		// get the email
-		$person = Connection::query("SELECT * FROM person WHERE username='$username'");
-
-		// return the email or false if not found
-		if(empty($person)) return false;
-		else return $person[0];
-	}
-
-	/**
-	 * Get the email from an id
-	 *
-	 * @author salvipascual
-	 * @param Int $id
-	 * @return String email or false
-	 */
-	public static function getEmailFromId($id)
-	{
-		// do not try empty inputs
-		if(empty($id)) return false;
-
-		// get the email
-		$email = Connection::query("SELECT email FROM person WHERE id=$id");
-
-		// return the email or false if not found
-		if(empty($email)) return false;
-		else return $email[0]->email;
 	}
 
 	/**
@@ -863,12 +826,10 @@ class Utils
 		$html = $doc->saveHTML();
 
 		$pbody = stripos($html, '<body>');
-		if ($pbody !== false)
-			$html = substr($html, $pbody + 6);
+		if ($pbody !== false) $html = substr($html, $pbody + 6);
 
 		$pbody = stripos($html, '</body');
-		if ($pbody !== false)
-			$html = substr($html, 0, $pbody);
+		if ($pbody !== false) $html = substr($html, 0, $pbody);
 
 		return $html;
 	}
@@ -878,39 +839,18 @@ class Utils
 	 *
 	 * @param string $path
 	 */
-	public static function rmdir($path){
+	public static function rmdir($path)
+	{
 		if (is_dir($path)) {
 			$dir = scandir($path);
-			foreach ( $dir as $d )
-			{
+			foreach ( $dir as $d ) {
 				if ($d != "." && $d != "..") {
-					if (is_dir("$path/$d"))
-					{
-						self::rmdir("$path/$d");
-					}
-					else
-					{
-						unlink("$path/$d");
-					}
+					if (is_dir("$path/$d")) self::rmdir("$path/$d");
+					else unlink("$path/$d");
 				}
 			}
 			rmdir($path);
 		}
-	}
-
-	/**
-	 * Get the completion percentage of a profile
-	 *
-	 * @REMOVE delete from the system and remove
-	 * @author salvipascual
-	 * @param String $email
-	 *
-	 * @return Number, percentage of completion
-	 */
-	public static function getProfileCompletion($email)
-	{
-		$profile = self::getPerson($email);
-		return $profile->completion;
 	}
 
 	/**
@@ -1033,8 +973,8 @@ class Utils
 	 * @param Response $response
 	 * @return array (Object, Attachments)
 	 */
-	public static function getAppData($person, $input, &$response){
-		
+	public static function getAppData($person, $input, &$response)
+	{	
 		// create the response
 		$appData = new stdClass();
 		$appData->reload = $input->command=="reload";
